feat: improve android implementation
